test(pdf): extract figure recovery tests
